答题-提交，其中涉及LastInsertId的问题，本来golang 自带的database/sql包可以完全解决的，但是这里用了gorm，就在事务里面用了Last解决这个问题

+++ models/topicModel.go
@@ -9,7 +9,7 @@ type TopicModel struct {
 	Type          string        `json:"type,omitempty"`
 	Options       string        `json:"options,omitempty"`
 	Explan        string        `json:"explan,omitempty"`
-	Score         int           `json:"score,omitempty"`
+	Score         int64         `json:"score,omitempty"`
 	Difficulty    string        `json:"difficulty,omitempty"`
 	Status        string        `json:"status,omitempty"`
 	ExtendType    string        `json:"extend_type,omitempty"`
